pyproject.toml
[build-system]
build-backend = "hatchling.build"

requires = [ "hatchling" ]

[project]
name = "djangocms-stories"
description = "The versatile stories and blog application for django CMS 4+"
readme = "README.rst"
license = "BSD-3-Clause"
authors = [
  { name = "Fabian Braun", email = "fsbraun@gmx.de" },
]
requires-python = ">=3.9"
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Framework :: Django",
  "Framework :: Django :: 4.2",
  "Framework :: Django :: 5.0",
  "Framework :: Django :: 5.1",
  "Framework :: Django :: 5.2",
  "Framework :: Django CMS :: 4.1",
  "Framework :: Django CMS :: 5.0",
  "License :: OSI Approved :: BSD License",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dynamic = [ "version" ]
dependencies = [
  "django-cms>=4.1",
  "django-filer>=2",
  "django-meta>=2",
  "django-parler>=2",
  "django-sortedm2m",
  "django-taggit>=1",
  "django-taggit-autosuggest",
  "django-taggit-templatetags",
  "easy-thumbnails>=2.4.1",
  "lxml",
  "pytz",
]

optional-dependencies.docs = [
  "django<5",
]
optional-dependencies.taggit-helpers = [
  "django-taggit-helpers",
]
optional-dependencies.test = [
  "coverage",
  "coveralls>=2",
  "django-taggit-helpers",
  "djangocms-text",
  "djangocms-video",
  "pysolr",
  "pytest",
  "pytest-asyncio",
  "pytest-django",
]
# urls.Documentation = "https://djangocms-blog.readthedocs.io/"
urls.Homepage = "https://github.com/django-cms/djangocms-stories"
# urls."Release notes" = ""

urls.Issues = "https://github.com/fsbraun/djangocms-stories/issues"

[tool.hatch.version]
path = "djangocms_stories/__init__.py"

[tool.hatch.build.targets.sdist]
include = [
  "/djangocms_blog",
  "/djangocms_stories",
]

[tool.black]
line-length = 119
target-version = [ "py39" ]

[tool.ruff]
target-version = "py39"
line-length = 119
exclude = [ "djangocms_blog", "djangocms_stories/migrations", "tests/fixtures" ]
lint.ignore = [  ]
lint.mccabe.max-complexity = 5

[tool.isort]
profile = "black"
combine_as_imports = true
default_section = "THIRDPARTY"
force_grid_wrap = 0
include_trailing_comma = true
known_first_party = "knocker"
line_length = 119
multi_line_output = 3
use_parentheses = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
addopts = [
  "--nomigrations",
]
python_files = [ "tests.py", "test_*.py" ]
testpaths = [ "tests" ]
pythonpath = [ "." ]

[tool.coverage.run]
source = [
  "djangocms_stories",
]
omit = [
  "djangocms_stories/migrations/*",
]
